+++ CHANGES.md
@@ -3,7 +3,8 @@
 
 ## Unreleased
 
-- Added `cratedb_toolkit.datasets` subsystem
+- Added `cratedb_toolkit.datasets` subsystem, for acquiring datasets
+  from cratedb-datasets and Kaggle.
 
 
 ## 2024/02/12 v0.0.5